haskellPackages: add hoogleLocal
The docs in `hoogle.nix` intend for it to be in the packageset as `hoogleLocal`
and that also makes a lot of sense from a user perspective.

`packages` is not a function `HaskellPackages -> [Package]`, but that is good,
because that way it can be used e.g. to merge package databases from other
versions of the packageset.