buildSupport: withTests nixpkgs testing base function
The idea is that you can test derivations without having to rebuild if a test
fails or needing a VM-test.

So
foo = withTests { t1 = ''exit 1''; } fooDrv
will build fooDrv, but fail the build of foo because test t1 exits ≠ 0.

This is the base building block for a nixpkgs testing framework with no nixos
dependency.