buildSupport/package-tests: abstract drvSeq from withTests
withTests is just a small wrapper around the more abstract functions `drvSeq`
and `drvSeqL`, which act as `seq` for derivations.
It might be possible to make them even more abstract by not requiring (some of)
their arguments to be derivations, but this is a nice first step.

Maybe the functions should move to `lib/trivial` instead?