lib/tests/local-nix-store: add tests
Preparation to test `withTests` and `drvSeq(L)`.
This already uses `withTests` to test the local storage code. Testception!

It is ensured that instantiation is actually done in a local store and
realization works as well. Also serves as a reminder for when nix syntax
changes (i.e. on the change to `1.12`).