shellcheck: add manpage and docs
The shellcheck source code contains a manpage, so let’s add that to
the output. For good measure, also copy the README to the doc output.

The overrideMeta thing is taking care of setting the right position,
so that `nix edit` points to this file instead of to the original
haskellPackages definition.