lib/licenses: mark deprecated licenses as deprecated and do meta check
We deprecated some gpl-like licenses because they were not unique
enough.
https://discourse.nixos.org/t/lib-licenses-gpl3-co-are-now-deprecated/8206

Add a deprecation field to these licenses and enrich the meta
typecheck so that packages that use the deprecated license will fail
the meta check the next time they are updated.

This should hopefully not influence normal evaluation, but only show
up when the ofborg meta check runs.